requirements.txt
click
sqlalchemy
sqlalchemy_utils
prettytable
pymysql
pyyaml
flask
